refactor: `PaymentApprovalProcessor` 분리 및 `PaymentServiceImpl` 의존성 주입
- 결제 승인 관련 로직을 `PaymentApprovalProcessor` 클래스로 분리하여 SRP 준수
- `PaymentServiceImpl`에서 `PaymentApprovalProcessor`를 활용하도록 코드 리팩토링
- `application-dev.yml`에 Hibernate 설정 추가 (`dialect`, 배치 크기 등)

ZBR-227
